requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-jose
passlib[bcrypt]
pytest
python-dotenv
